ZipProject/backend/requirements.txt
Django==3.2.6
django-allauth==0.45.0
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-seed==0.2.2
django-storages==1.11.1
djangorestframework==3.12.4
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
boto3
gunicorn==20.0.4